requirements.txt
flask 
numpy
gunicorn
dnspython
firebase-admin
python-dotenv